Detect pixel density on android and apply it
This problem only occurs on android, ios handle scalability
automatically. To behave correctly all point information need to use the
calaosApp.density information.
Another problem not handled in this commit is *.png files. On ios qt
automatically take @2x.png files. But it's not the case on android. I
need to write something to do it automatically. (by writing a special
image provider and override the default qrc one)