Ignore errors from http when using new features
On v1 (maybe v2 too) sending a non existing action will trigger a HTTP
forbidden error code. This makes our app close the connection because of
that error code. To prevent this, we simply ignore the http errors when
sending those actions. They are not supported by v1/v2 and are used
through websocket on v3 anyway.